+++ OrdersMS/OrdersMS/src/Orders/Application/Types/GetOrderResponse.cs
@@ -8,7 +8,7 @@
     CoordinatesDto DestinationAddress,
     string IncidentDate,
     List<ExtraServiceDto> ExtraServicesApplied,
-    double TotalCost,
+    decimal TotalCost,
     string Status
 );
